each components seperated
