+++ src/components/home/HomePage.jsx
@@ -6,6 +6,8 @@ import Newsletterbox from '../other/Newsletterbox';
 import ProductSlider from '../products/ProductSlider';
 import Ourpolicies from '../details/Ourpolicies';
 import TopSales from './TopSales';
+import JustForYou from './JustForYou'; 
+import ShopByCategory from './ShopByCategory';
 
 const backendUrl = import.meta.env.VITE_BACKEND_URL || 'http://localhost:5174';
 
@@ -133,104 +135,19 @@ const HomePage = ({ addToCart }) => {
       <TopSales addToCart={addToCart} />
 
       {/* Shop by Category Section */}
-      <div className="mt-10 mb-6 px-4 md:px-10">
-        <h2 className="text-2xl font-bold mb-4">Shop by Category</h2>
-        <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-6 gap-6">
-          {categories.slice(0, 6).map((category) => {
-            const latestProduct = items.find(item => item.category === category);
-            return (
-              <div key={category} className="bg-white p-4 rounded-lg shadow-md">
-                <h3 className="font-bold text-lg mb-2">{category}</h3>
-                {latestProduct ? (
-                  <div
-                    className="cursor-pointer"
-                    onClick={() => handleCategoryClick(category)}
-                  >
-                    <img
-                      src={latestProduct.photos[0]?.url || assets.placeholder}
-                      alt={latestProduct.title}
-                      className="w-full h-32 object-cover rounded-lg mb-2"
-                    />
-                  </div>
-                ) : (
-                  <div className="w-full h-32 bg-gray-200 rounded-lg mb-2 flex items-center justify-center">
-                    <span className="text-gray-500">No Products</span>
-                  </div>
-                )}
-                <Link
-                  to={`/category/${category}`}
-                  className="inline-block mt-3 text-indigo-600 text-sm font-semibold hover:underline"
-                >
-                  See More
-                </Link>
-              </div>
-            );
-          })}
-        </div>
-        {categories.length > 6 && (
-          <div className="flex justify-center mt-4">
-            <Link
-              to="/collection"
-              className="px-4 py-1 bg-indigo-600 text-white rounded-lg shadow-md text-sm hover:bg-indigo-500 transition-colors"
-            >
-              Explore More Categories
-            </Link>
-          </div>
-        )}
-      </div>
+      <ShopByCategory 
+        categories={categories} 
+        items={items} 
+        handleCategoryClick={handleCategoryClick} 
+      />
       {/* Just for You Section */}
-      <div className="px-4 md:px-10">
-        <h2 className="text-2xl font-bold mb-4">Just for You</h2>
-        {loading ? (
-          <div className="flex justify-center my-6">
-            <p className="text-gray-600">Loading products...</p>
-          </div>
-        ) : (
-          <>
-            <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 gap-4">
-              {items.slice(0, 10).map((product) => (
-                <div 
-                  key={product.id} 
-                  className="bg-white shadow-lg rounded-lg overflow-hidden cursor-pointer hover:shadow-xl transition-shadow duration-300 transform hover:scale-105"
-                  onClick={() => handleProductClick(product)}
-                >
-                  <ProductSlider images={product.photos.map(photo => photo.url)} />
-                  <div className="p-4">
-                    <h3 className="font-bold text-lg mb-2">{product.title}</h3>
-                    <p className="text-gray-600 text-sm mb-2">{product.description}</p>
-                    <p className="text-gray-800 font-semibold mb-2">Price: ${product.price}</p>
-                    
-                    {/* Rating Display */}
-                    {product.rating && renderStars(product.rating)}
-      
-                    {/* Stock Display with Conditional Styling */}
-                    <p className={`text-sm font-semibold ${product.inStock > 5 ? 'text-green-600' : 'text-red-600'}`}>
-                      {product.inStock > 5 ? `In Stock: ${product.inStock}` : 'Low Stock'}
-                    </p>
-      
-                    <button
-                      onClick={(e) => { e.stopPropagation(); handleAddToCart(product); }}
-                      className="mt-2 w-full bg-indigo-600 text-white py-2 rounded-lg font-medium hover:bg-indigo-500 transition-colors"
-                    >
-                      Add to Cart
-                    </button>
-                  </div>
-                </div>
-              ))}
-            </div>
-            {items.length > 10 && (
-              <div className="flex justify-center mt-6">
-                <Link
-                  to="/collection"
-                  className="px-6 py-2 bg-indigo-600 text-white rounded-lg shadow-md text-sm font-semibold hover:bg-indigo-500 transition-colors"
-                >
-                  See More Products
-                </Link>
-              </div>
-            )}
-          </>
-        )}
-      </div>
+      <JustForYou 
+        items={items} 
+        loading={loading} 
+        handleProductClick={handleProductClick} 
+        handleAddToCart={handleAddToCart} 
+        renderStars={renderStars} 
+      />
       <Ourpolicies />
       {!showNewsletterPopup && <Newsletterbox mode="inline" />}
       {showModal && <CartModal closeModal={() => setShowModal(false)} />}
